+++ src/components.d.ts
@@ -12,10 +12,12 @@ import { IAccountScreenData, IConfirmScreenData, IConnectScreenData, ILedgerConn
 import { IEventBus } from "./utils/EventBus";
 import { IPendingTransactionsModalData } from "./components/functional/pending-transactions-modal/pending-transactions-modal.types";
 import { ProviderTypeEnum } from "./types/provider.types";
+import { SidePanelSideEnum } from "./components/visual/side-panel/side-panel.types";
 import { LocalJSX as JSX, VNode } from "@stencil/core";
 import { ISignTransactionsModalData } from "./components/functional/sign-transactions-modal/sign-transactions-modal.types";
 import { CustomToastType as CustomToastType1, IToastDataState, ITransaction, ITransactionProgressState, ITransactionToast } from "./components/functional/toasts-list/components/transaction-toast/transaction-toast.type";
 import { ITransactionAccount, ITransactionIconInfo, ITransactionsTableRow } from "./components/controlled/transactions-table/transactions-table.type";
+import { ITransactionListItem } from "./components/visual/transaction-list-item/transaction-list-item.types";
 import { ITransactionValue } from "./components/controlled/transactions-table/transactions-table.type";
 import { IWalletConnectModalData } from "./components/functional/wallet-connect-components/wallet-connect-modal.types";
 export { IconDefinition } from "@fortawesome/free-solid-svg-icons";
@@ -25,10 +27,12 @@ export { IAccountScreenData, IConfirmScreenData, IConnectScreenData, ILedgerConn
 export { IEventBus } from "./utils/EventBus";
 export { IPendingTransactionsModalData } from "./components/functional/pending-transactions-modal/pending-transactions-modal.types";
 export { ProviderTypeEnum } from "./types/provider.types";
+export { SidePanelSideEnum } from "./components/visual/side-panel/side-panel.types";
 export { LocalJSX as JSX, VNode } from "@stencil/core";
 export { ISignTransactionsModalData } from "./components/functional/sign-transactions-modal/sign-transactions-modal.types";
 export { CustomToastType as CustomToastType1, IToastDataState, ITransaction, ITransactionProgressState, ITransactionToast } from "./components/functional/toasts-list/components/transaction-toast/transaction-toast.type";
 export { ITransactionAccount, ITransactionIconInfo, ITransactionsTableRow } from "./components/controlled/transactions-table/transactions-table.type";
+export { ITransactionListItem } from "./components/visual/transaction-list-item/transaction-list-item.types";
 export { ITransactionValue } from "./components/controlled/transactions-table/transactions-table.type";
 export { IWalletConnectModalData } from "./components/functional/wallet-connect-components/wallet-connect-modal.types";
 export namespace Components {
@@ -57,7 +61,7 @@ export namespace Components {
     interface FaIcon {
         "class"?: string;
         "description"?: string;
-        "icon": IconDefinition;
+        "icon": IconDefinition | string;
     }
     interface FormatAmount {
         "class"?: string;
@@ -99,12 +103,20 @@ export namespace Components {
     interface LedgerConnectScreen {
         "connectScreenData": IConnectScreenData;
     }
+    interface NotificationsFeed {
+        "getEventBus": () => Promise<IEventBus>;
+    }
     interface PendingTransactionsModal {
         "data": IPendingTransactionsModalData;
         "getEventBus": () => Promise<IEventBus>;
     }
     interface ProviderButton {
         "type": ProviderTypeEnum;
+    }
+    interface SidePanel {
+        "isOpen": boolean;
+        "panelClassName"?: string;
+        "side": SidePanelSideEnum;
     }
     interface SignTransactionComponent {
         "header": VNode;
@@ -153,6 +165,9 @@ export namespace Components {
     interface TransactionIcon {
         "class"?: string;
         "iconInfo": ITransactionIconInfo;
+    }
+    interface TransactionListItem {
+        "transaction": ITransactionListItem;
     }
     interface TransactionMethod {
         "actionDescription": string;
@@ -220,7 +235,7 @@ export namespace Components {
     }
     interface UnlockPanel {
         "allowedProviders"?: ProviderTypeEnum[];
-        "open": boolean;
+        "isOpen": boolean;
     }
     interface WalletConnectModal {
         "data": IWalletConnectModalData;
@@ -246,6 +261,10 @@ export interface LedgerAccountScreenCustomEvent<T> extends CustomEvent<T> {
 export interface LedgerConnectScreenCustomEvent<T> extends CustomEvent<T> {
     detail: T;
     target: HTMLLedgerConnectScreenElement;
+}
+export interface SidePanelCustomEvent<T> extends CustomEvent<T> {
+    detail: T;
+    target: HTMLSidePanelElement;
 }
 export interface SimpleToastCustomEvent<T> extends CustomEvent<T> {
     detail: T;
@@ -416,6 +435,12 @@ declare global {
         prototype: HTMLLedgerConnectScreenElement;
         new (): HTMLLedgerConnectScreenElement;
     };
+    interface HTMLNotificationsFeedElement extends Components.NotificationsFeed, HTMLStencilElement {
+    }
+    var HTMLNotificationsFeedElement: {
+        prototype: HTMLNotificationsFeedElement;
+        new (): HTMLNotificationsFeedElement;
+    };
     interface HTMLPendingTransactionsModalElement extends Components.PendingTransactionsModal, HTMLStencilElement {
     }
     var HTMLPendingTransactionsModalElement: {
@@ -427,6 +452,23 @@ declare global {
     var HTMLProviderButtonElement: {
         prototype: HTMLProviderButtonElement;
         new (): HTMLProviderButtonElement;
+    };
+    interface HTMLSidePanelElementEventMap {
+        "close": any;
+    }
+    interface HTMLSidePanelElement extends Components.SidePanel, HTMLStencilElement {
+        addEventListener<K extends keyof HTMLSidePanelElementEventMap>(type: K, listener: (this: HTMLSidePanelElement, ev: SidePanelCustomEvent<HTMLSidePanelElementEventMap[K]>) => any, options?: boolean | AddEventListenerOptions): void;
+        addEventListener<K extends keyof DocumentEventMap>(type: K, listener: (this: Document, ev: DocumentEventMap[K]) => any, options?: boolean | AddEventListenerOptions): void;
+        addEventListener<K extends keyof HTMLElementEventMap>(type: K, listener: (this: HTMLElement, ev: HTMLElementEventMap[K]) => any, options?: boolean | AddEventListenerOptions): void;
+        addEventListener(type: string, listener: EventListenerOrEventListenerObject, options?: boolean | AddEventListenerOptions): void;
+        removeEventListener<K extends keyof HTMLSidePanelElementEventMap>(type: K, listener: (this: HTMLSidePanelElement, ev: SidePanelCustomEvent<HTMLSidePanelElementEventMap[K]>) => any, options?: boolean | EventListenerOptions): void;
+        removeEventListener<K extends keyof DocumentEventMap>(type: K, listener: (this: Document, ev: DocumentEventMap[K]) => any, options?: boolean | EventListenerOptions): void;
+        removeEventListener<K extends keyof HTMLElementEventMap>(type: K, listener: (this: HTMLElement, ev: HTMLElementEventMap[K]) => any, options?: boolean | EventListenerOptions): void;
+        removeEventListener(type: string, listener: EventListenerOrEventListenerObject, options?: boolean | EventListenerOptions): void;
+    }
+    var HTMLSidePanelElement: {
+        prototype: HTMLSidePanelElement;
+        new (): HTMLSidePanelElement;
     };
     interface HTMLSignTransactionComponentElement extends Components.SignTransactionComponent, HTMLStencilElement {
     }
@@ -504,6 +546,12 @@ declare global {
     var HTMLTransactionIconElement: {
         prototype: HTMLTransactionIconElement;
         new (): HTMLTransactionIconElement;
+    };
+    interface HTMLTransactionListItemElement extends Components.TransactionListItem, HTMLStencilElement {
+    }
+    var HTMLTransactionListItemElement: {
+        prototype: HTMLTransactionListItemElement;
+        new (): HTMLTransactionListItemElement;
     };
     interface HTMLTransactionMethodElement extends Components.TransactionMethod, HTMLStencilElement {
     }
@@ -657,8 +705,10 @@ declare global {
         "ledger-connect": HTMLLedgerConnectElement;
         "ledger-connect-modal": HTMLLedgerConnectModalElement;
         "ledger-connect-screen": HTMLLedgerConnectScreenElement;
+        "notifications-feed": HTMLNotificationsFeedElement;
         "pending-transactions-modal": HTMLPendingTransactionsModalElement;
         "provider-button": HTMLProviderButtonElement;
+        "side-panel": HTMLSidePanelElement;
         "sign-transaction-component": HTMLSignTransactionComponentElement;
         "sign-transactions-modal": HTMLSignTransactionsModalElement;
         "simple-toast": HTMLSimpleToastElement;
@@ -670,6 +720,7 @@ declare global {
         "transaction-direction-badge": HTMLTransactionDirectionBadgeElement;
         "transaction-hash": HTMLTransactionHashElement;
         "transaction-icon": HTMLTransactionIconElement;
+        "transaction-list-item": HTMLTransactionListItemElement;
         "transaction-method": HTMLTransactionMethodElement;
         "transaction-shards": HTMLTransactionShardsElement;
         "transaction-toast": HTMLTransactionToastElement;
@@ -714,7 +765,7 @@ declare namespace LocalJSX {
     interface FaIcon {
         "class"?: string;
         "description"?: string;
-        "icon"?: IconDefinition;
+        "icon"?: IconDefinition | string;
     }
     interface FormatAmount {
         "class"?: string;
@@ -761,11 +812,19 @@ declare namespace LocalJSX {
         "connectScreenData"?: IConnectScreenData;
         "onConnect"?: (event: LedgerConnectScreenCustomEvent<any>) => void;
     }
+    interface NotificationsFeed {
+    }
     interface PendingTransactionsModal {
         "data"?: IPendingTransactionsModalData;
     }
     interface ProviderButton {
         "type"?: ProviderTypeEnum;
+    }
+    interface SidePanel {
+        "isOpen"?: boolean;
+        "onClose"?: (event: SidePanelCustomEvent<any>) => void;
+        "panelClassName"?: string;
+        "side"?: SidePanelSideEnum;
     }
     interface SignTransactionComponent {
         "header"?: VNode;
@@ -813,6 +872,9 @@ declare namespace LocalJSX {
     interface TransactionIcon {
         "class"?: string;
         "iconInfo"?: ITransactionIconInfo;
+    }
+    interface TransactionListItem {
+        "transaction"?: ITransactionListItem;
     }
     interface TransactionMethod {
         "actionDescription"?: string;
@@ -884,9 +946,9 @@ declare namespace LocalJSX {
     }
     interface UnlockPanel {
         "allowedProviders"?: ProviderTypeEnum[];
+        "isOpen"?: boolean;
         "onClose"?: (event: UnlockPanelCustomEvent<any>) => void;
         "onLogin"?: (event: UnlockPanelCustomEvent<{ provider: ProviderTypeEnum; anchor?: HTMLElement }>) => void;
-        "open"?: boolean;
     }
     interface WalletConnectModal {
         "data"?: IWalletConnectModalData;
@@ -907,8 +969,10 @@ declare namespace LocalJSX {
         "ledger-connect": LedgerConnect;
         "ledger-connect-modal": LedgerConnectModal;
         "ledger-connect-screen": LedgerConnectScreen;
+        "notifications-feed": NotificationsFeed;
         "pending-transactions-modal": PendingTransactionsModal;
         "provider-button": ProviderButton;
+        "side-panel": SidePanel;
         "sign-transaction-component": SignTransactionComponent;
         "sign-transactions-modal": SignTransactionsModal;
         "simple-toast": SimpleToast;
@@ -920,6 +984,7 @@ declare namespace LocalJSX {
         "transaction-direction-badge": TransactionDirectionBadge;
         "transaction-hash": TransactionHash;
         "transaction-icon": TransactionIcon;
+        "transaction-list-item": TransactionListItem;
         "transaction-method": TransactionMethod;
         "transaction-shards": TransactionShards;
         "transaction-toast": TransactionToast;
@@ -956,8 +1021,10 @@ declare module "@stencil/core" {
             "ledger-connect": LocalJSX.LedgerConnect & JSXBase.HTMLAttributes<HTMLLedgerConnectElement>;
             "ledger-connect-modal": LocalJSX.LedgerConnectModal & JSXBase.HTMLAttributes<HTMLLedgerConnectModalElement>;
             "ledger-connect-screen": LocalJSX.LedgerConnectScreen & JSXBase.HTMLAttributes<HTMLLedgerConnectScreenElement>;
+            "notifications-feed": LocalJSX.NotificationsFeed & JSXBase.HTMLAttributes<HTMLNotificationsFeedElement>;
             "pending-transactions-modal": LocalJSX.PendingTransactionsModal & JSXBase.HTMLAttributes<HTMLPendingTransactionsModalElement>;
             "provider-button": LocalJSX.ProviderButton & JSXBase.HTMLAttributes<HTMLProviderButtonElement>;
+            "side-panel": LocalJSX.SidePanel & JSXBase.HTMLAttributes<HTMLSidePanelElement>;
             "sign-transaction-component": LocalJSX.SignTransactionComponent & JSXBase.HTMLAttributes<HTMLSignTransactionComponentElement>;
             "sign-transactions-modal": LocalJSX.SignTransactionsModal & JSXBase.HTMLAttributes<HTMLSignTransactionsModalElement>;
             "simple-toast": LocalJSX.SimpleToast & JSXBase.HTMLAttributes<HTMLSimpleToastElement>;
@@ -969,6 +1036,7 @@ declare module "@stencil/core" {
             "transaction-direction-badge": LocalJSX.TransactionDirectionBadge & JSXBase.HTMLAttributes<HTMLTransactionDirectionBadgeElement>;
             "transaction-hash": LocalJSX.TransactionHash & JSXBase.HTMLAttributes<HTMLTransactionHashElement>;
             "transaction-icon": LocalJSX.TransactionIcon & JSXBase.HTMLAttributes<HTMLTransactionIconElement>;
+            "transaction-list-item": LocalJSX.TransactionListItem & JSXBase.HTMLAttributes<HTMLTransactionListItemElement>;
             "transaction-method": LocalJSX.TransactionMethod & JSXBase.HTMLAttributes<HTMLTransactionMethodElement>;
             "transaction-shards": LocalJSX.TransactionShards & JSXBase.HTMLAttributes<HTMLTransactionShardsElement>;
             "transaction-toast": LocalJSX.TransactionToast & JSXBase.HTMLAttributes<HTMLTransactionToastElement>;