+++ src/components/functional/toasts-list/toast-list.tsx
@@ -4,6 +4,7 @@ import { EventBus } from 'utils/EventBus';
 
 import type { CustomToastType, ITransactionToast } from './components/transaction-toast/transaction-toast.type';
 import { ToastEventsEnum } from './toast-list.types';
+
 @Component({
   tag: 'toast-list',
   styleUrl: 'toast-list.css',
@@ -17,6 +18,11 @@ export class ToastList {
     return this.eventBus;
   }
 
+  disconnectedCallback() {
+    this.eventBus.unsubscribe(ToastEventsEnum.TRANSACTION_TOAST_DATA_UPDATE, this.transactionToastUpdate.bind(this));
+    this.eventBus.unsubscribe(ToastEventsEnum.CUSTOM_TOAST_DATA_UPDATE, this.customToastsUpdate.bind(this));
+  }
+
   private handleCustomToastDelete(toastId: string) {
     this.eventBus.publish(ToastEventsEnum.CLOSE_TOAST, toastId);
   }
@@ -25,11 +31,22 @@ export class ToastList {
     this.eventBus.publish(ToastEventsEnum.CLOSE_TOAST, toastId);
   }
 
+  private handleViewAllClick() {
+    this.eventBus.publish(ToastEventsEnum.OPEN_NOTIFICATIONS_FEED);
+  }
+
   render() {
     return (
       <div class="toast-list" id="toast-list">
         {this.customToasts?.map(toast => <generic-toast toast={toast} onHandleDeleteToast={this.handleCustomToastDelete.bind(this, toast.toastId)}></generic-toast>)}
         {this.transactionToasts?.map(toast => <transaction-toast {...toast} onHandleDeleteToast={this.handleTransactionToastDelete.bind(this, toast.toastId)}></transaction-toast>)}
+        {this.transactionToasts?.length > 0 && (
+          <div class="view-all-button-container">
+            <button class="view-all-button" onClick={this.handleViewAllClick.bind(this)}>
+              View All
+            </button>
+          </div>
+        )}
       </div>
     );
   }